Update policies
Add some more best practices that I've seen come up on reviews quite often.

Add a section on not returning null since the canonical data often cause
confusion about this.

Add a section on using ExpectedException since this was something we agreed
a while ago but it's not documented anywhere.