Add test case for Bird Watcher exercise (#2956)
Add a test to `hasDayWithoutBirds` for when the last day has zero visits. This checks that the solution correctly checks the entire array. See also a the forum post https://forum.exercism.org/t/bird-watcher-in-java-should-we-add-a-test-case/17933/16.

Also make `lastWeek` final in the test, since the field isn't meant to change in the test.

Co-authored-by: Jegors Cemisovs <jcemisovs@evolution.com>